--- src/components/Bookmarks.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import './Bookmarks.css';
-
-type Props = {};
-
-const Bookmarks = (props: Props) => {
-  return <div>Bookmarks</div>;
-};
-
-export default Bookmarks;
