+++ src/components/MangaDetails.tsx
@@ -4,7 +4,11 @@ import './MangaDetails.css';
 type Props = {};
 
 const MangaDetails = (props: Props) => {
-  return <div>MangaDetails</div>;
+  return (
+    <div className="details">
+      <h3>MangaDetails cared</h3>
+    </div>
+  );
 };
 
 export default MangaDetails;
